+++ tests/Unit/RemoteConfig/DefaultValueTest.php
@@ -5,16 +5,22 @@ declare(strict_types=1);
 namespace Kreait\Firebase\Tests\Unit\RemoteConfig;
 
 use Kreait\Firebase\RemoteConfig\DefaultValue;
+use Kreait\Firebase\RemoteConfig\ExplicitValue;
+use Kreait\Firebase\RemoteConfig\PersonalizationValue;
 use PHPUnit\Framework\TestCase;
 
 /**
  * @internal
+ *
+ * @phpstan-import-type RemoteConfigPersonalizationValueShape from PersonalizationValue
+ * @phpstan-import-type RemoteConfigExplicitValueShape from ExplicitValue
+ * @phpstan-import-type RemoteConfigInAppDefaultValueShape from DefaultValue
  */
 final class DefaultValueTest extends TestCase
 {
     public function testCreateInAppDefaultValue(): void
     {
-        $defaultValue = DefaultValue::none();
+        $defaultValue = DefaultValue::useInAppDefault();
 
         $this->assertTrue($defaultValue->value());
         $this->assertEquals(['useInAppDefault' => true], $defaultValue->jsonSerialize());
@@ -31,29 +37,34 @@ final class DefaultValueTest extends TestCase
     /**
      * @dataProvider arrayValueProvider
      *
-     * @param bool|string $expected
-     * @param array{
-     *     value: string|bool
-     * }|array{
-     *     useInAppDefault: bool
-     * } $data
+     * @param RemoteConfigInAppDefaultValueShape $expected
+     * @param RemoteConfigInAppDefaultValueShape $data
      */
-    public function testCreateFromArray($expected, array $data): void
+    public function testCreateFromArray(array $expected, array $data): void
     {
         $defaultValue = DefaultValue::fromArray($data);
 
-        $this->assertSame($expected, $defaultValue->value());
+        $this->assertSame($expected, $defaultValue->toArray());
     }
 
     /**
-     * @return iterable<array{value?: string|bool, useInAppDefault?: bool}>
+     * @return iterable<non-empty-string, array<RemoteConfigInAppDefaultValueShape|RemoteConfigExplicitValueShape|RemoteConfigPersonalizationValueShape>>
      */
-    public function arrayValueProvider()
+    public function arrayValueProvider(): iterable
     {
-        yield 'inAppDefault' => [true, ['useInAppDefault' => true]];
+        yield 'inAppDefault' => [
+            ['useInAppDefault' => true],
+            ['useInAppDefault' => true],
+        ];
 
-        yield 'bool' => [true, ['value' => true]];
+        yield 'explicit' => [
+            ['value' => '1'],
+            ['value' => '1'],
+        ];
 
-        yield 'string' => ['foo', ['value' => 'foo']];
+        yield 'personalization' => [
+            ['personalizationId' => 'pid'],
+            ['personalizationId' => 'pid'],
+        ];
     }
 }